try to fix various ways ansi-color gets confused by repl, use newer advice api
try to fix `error in process filter: Invalid search bound (wrong side of point)`
also try to avoid some cases where it stops interpreting colors

May still be some remaining performance problem somewhere? Sometimes
after clearing repl, that repl is super slow forever (seconds to notice
input, or to eval/print commands, etc). When that happens, only workaround
I've found is to kill the repl buffer and reconnect to the running lisp.
Possibly that only happens when ansi-color is told to leave the escape
sequences in the buffer, so disabled that. Not sure if keeping them is
actually useful for anything besides debugging colors anyway though.